apps/minimal/pyproject.toml
[project]
name = "minimal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.5"
dependencies = [
    "matplotlib>=3.10.3",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
    "streamlit>=1.45.1",
]
